requirements.txt
python-telegram-bot==21.5
aiohttp==3.10.5
python-dotenv==1.0.0
aiosqlite==0.19.0